feat: extract-shared-app-dashboard-test-harness-001 - Share the App dashboard harness setup
