+++ ui/src/App.current-selection.test.tsx
@@ -1,93 +1,28 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import {
   act,
-  cleanup,
   fireEvent,
-  render,
   screen,
   waitFor,
   within,
 } from "@testing-library/react";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { App } from "./App";
+import { describe, expect, it } from "vitest";
+import {
+  MockEventSource,
+  registerAppDashboardTestLifecycle,
+  removeTraceIDsFromSnapshot,
+  renderApp,
+  requireValue,
+} from "./App.dashboard-test-harness";
 import type {
   DashboardSnapshot,
   DashboardTrace,
-  DashboardWorkItemRef,
-  DashboardWorkstationRequest,
 } from "./api/dashboard";
 import {
   buildDashboardSnapshotFixture,
   dashboardWorkstationRequestFixtures,
   mediumBranchingDashboardTopology,
 } from "./components/dashboard/fixtures";
-import { installDashboardBrowserTestShims } from "./components/dashboard/test-browser-shims";
 import { semanticWorkflowDashboardSnapshot } from "./components/dashboard/test-fixtures";
-import { useDashboardBentoStore } from "./features/bento/state/dashboardBentoStore";
-import { reloadDashboardLayoutFromStorage } from "./features/bento/useDashboardLayout";
-import { useCurrentEditableFactoryDefinition } from "./features/current-factory-definition";
-import { resetSelectionHistoryStore } from "./features/current-selection/state/selectionHistoryStore";
-import {
-  createDefaultDashboardStreamState,
-  useDashboardStreamStore,
-} from "./features/dashboard/state/dashboardStreamStore";
-import { useExportDialogStore } from "./features/export/state/exportDialogStore";
-import type { WorldState } from "./features/timeline/state/factoryTimelineStore";
-import { useFactoryTimelineStore } from "./features/timeline/state/factoryTimelineStore";
-
-vi.mock("./features/current-factory-definition", async () => {
-  const actual = await vi.importActual("./features/current-factory-definition");
-
-  return {
-    ...actual,
-    useCurrentEditableFactoryDefinition: vi.fn(),
-  };
-});
-
-class MockEventSource {
-  public static instances: MockEventSource[] = [];
-
-  public onerror: ((event: Event) => void) | null = null;
-  public onopen: ((event: Event) => void) | null = null;
-
-  private readonly listeners = new Map<string, EventListener[]>();
-
-  constructor(public readonly url: string) {
-    MockEventSource.instances.push(this);
-  }
-
-  public addEventListener(type: string, listener: EventListener): void {
-    const existing = this.listeners.get(type) ?? [];
-    existing.push(listener);
-    this.listeners.set(type, existing);
-  }
-
-  public close(): void {}
-
-  public emit(type: string, data: unknown): void {
-    if (type === "snapshot") {
-      const state = useFactoryTimelineStore.getState();
-      const tracesByWorkID =
-        state.worldViewCache[state.selectedTick]?.tracesByWorkID ?? {};
-      seedTimelineSnapshot(data as DashboardSnapshot, tracesByWorkID);
-    }
-
-    const event = new MessageEvent(type, {
-      data: JSON.stringify(data),
-    });
-
-    for (const listener of this.listeners.get(type) ?? []) {
-      listener(event);
-    }
-  }
-}
-
-interface RenderAppOptions {
-  snapshot: DashboardSnapshot;
-  timelineSnapshots?: DashboardSnapshot[];
-  traceFixtures?: Record<string, DashboardTrace>;
-  workstationRequestsByDispatchID?: Record<string, DashboardWorkstationRequest>;
-}
 
 const activeWorkID = "work-active-story";
 const completedWorkID = "work-complete";
@@ -272,133 +207,6 @@ const failedTraceSnapshot: DashboardTrace = {
   ],
 };
 
-const queryClients: QueryClient[] = [];
-let restoreBrowserTestShims: (() => void) | null = null;
-
-function timelineSnapshot(
-  snapshot: DashboardSnapshot,
-  tracesByWorkID: Record<string, DashboardTrace> = {},
-  workstationRequestsByDispatchID: Record<
-    string,
-    DashboardWorkstationRequest
-  > = {},
-): WorldState {
-  return {
-    ...snapshot,
-    relationsByWorkID: {},
-    tracesByWorkID,
-    workstationRequestsByDispatchID,
-    workRequestsByID: {},
-  };
-}
-
-function seedTimelineSnapshot(
-  snapshot: DashboardSnapshot,
-  tracesByWorkID: Record<string, DashboardTrace> = {},
-  workstationRequestsByDispatchID: Record<
-    string,
-    DashboardWorkstationRequest
-  > = {},
-): void {
-  useFactoryTimelineStore.setState({
-    events: [],
-    latestTick: snapshot.tick_count,
-    mode: "current",
-    receivedEventIDs: [],
-    selectedTick: snapshot.tick_count,
-    worldViewCache: {
-      [snapshot.tick_count]: timelineSnapshot(
-        snapshot,
-        tracesByWorkID,
-        workstationRequestsByDispatchID,
-      ),
-    },
-  });
-}
-
-function seedTimelineSnapshots(snapshots: DashboardSnapshot[]): void {
-  const worldViewCache = Object.fromEntries(
-    snapshots.map(
-      (snapshot) =>
-        [
-          snapshot.tick_count,
-          timelineSnapshot(snapshot) satisfies WorldState,
-        ] as const,
-    ),
-  );
-  const latestTick = Math.max(
-    ...snapshots.map((snapshot) => snapshot.tick_count),
-  );
-
-  useFactoryTimelineStore.setState({
-    events: [],
-    latestTick,
-    mode: "current",
-    receivedEventIDs: [],
-    selectedTick: latestTick,
-    worldViewCache,
-  });
-}
-
-function renderApp({
-  snapshot,
-  timelineSnapshots,
-  traceFixtures = {},
-  workstationRequestsByDispatchID = {},
-}: RenderAppOptions) {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        gcTime: Infinity,
-        retry: false,
-      },
-    },
-  });
-  queryClients.push(queryClient);
-
-  const fetchMock = vi
-    .fn()
-    .mockImplementation(async (input: RequestInfo | URL) => {
-      const path =
-        typeof input === "string"
-          ? input
-          : input instanceof URL
-            ? `${input.pathname}${input.search}`
-            : input.url;
-
-      throw new Error(`unexpected fetch for ${path}`);
-    });
-
-  vi.stubGlobal("fetch", fetchMock);
-  vi.stubGlobal("EventSource", MockEventSource);
-  reloadDashboardLayoutFromStorage();
-  if (timelineSnapshots) {
-    seedTimelineSnapshots(timelineSnapshots);
-  } else {
-    seedTimelineSnapshot(
-      snapshot,
-      traceFixtures,
-      workstationRequestsByDispatchID,
-    );
-  }
-
-  const result = render(
-    <QueryClientProvider client={queryClient}>
-      <App />
-    </QueryClientProvider>,
-  );
-
-  return { ...result, fetchMock };
-}
-
-function requireValue<T>(value: T | null | undefined, message: string): T {
-  if (value === null || value === undefined) {
-    throw new Error(message);
-  }
-
-  return value;
-}
-
 function getDispatchHistoryCard(
   container: HTMLElement,
   dispatchId: string,
@@ -465,73 +273,6 @@ function getActiveStorySelectionButton(): HTMLElement {
   return activeStoryButton;
 }
 
-function removeTraceIDFromWorkItem(
-  workItem: DashboardWorkItemRef,
-): DashboardWorkItemRef {
-  const withoutTraceID: DashboardWorkItemRef = { work_id: workItem.work_id };
-  if (workItem.display_name) {
-    withoutTraceID.display_name = workItem.display_name;
-  }
-  if (workItem.work_type_id) {
-    withoutTraceID.work_type_id = workItem.work_type_id;
-  }
-  return withoutTraceID;
-}
-
-function removeTraceIDsFromSnapshot(
-  snapshot: DashboardSnapshot,
-): DashboardSnapshot {
-  return {
-    ...snapshot,
-    runtime: {
-      ...snapshot.runtime,
-      active_executions_by_dispatch_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.active_executions_by_dispatch_id ?? {},
-        ).map(([dispatchID, execution]) => [
-          dispatchID,
-          {
-            ...execution,
-            trace_ids: [],
-            work_items: execution.work_items?.map(removeTraceIDFromWorkItem),
-          },
-        ]),
-      ),
-      current_work_items_by_place_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.current_work_items_by_place_id ?? {},
-        ).map(([placeID, workItems]) => [
-          placeID,
-          workItems.map(removeTraceIDFromWorkItem),
-        ]),
-      ),
-      session: {
-        ...snapshot.runtime.session,
-        provider_sessions: snapshot.runtime.session.provider_sessions?.map(
-          (attempt) => ({
-            ...attempt,
-            work_items: attempt.work_items?.map(removeTraceIDFromWorkItem),
-          }),
-        ),
-      },
-      workstation_activity_by_node_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.workstation_activity_by_node_id ?? {},
-        ).map(([nodeID, activity]) => [
-          nodeID,
-          {
-            ...activity,
-            active_work_items: activity.active_work_items?.map(
-              removeTraceIDFromWorkItem,
-            ),
-            trace_ids: [],
-          },
-        ]),
-      ),
-    },
-  };
-}
-
 function resizeDashboardViewport(width: number): void {
   Object.defineProperty(window, "innerWidth", {
     configurable: true,
@@ -547,59 +288,7 @@ function resizeDashboardViewport(width: number): void {
 }
 
 describe("App current selection", () => {
-  beforeEach(() => {
-    window.localStorage.clear();
-    MockEventSource.instances = [];
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-    resetSelectionHistoryStore();
-    vi.mocked(useCurrentEditableFactoryDefinition).mockReturnValue({
-      data: undefined,
-      error: null,
-      failureCount: 0,
-      failureReason: null,
-      fetchStatus: "idle",
-      isError: false,
-      isFetched: false,
-      isFetchedAfterMount: false,
-      isFetching: false,
-      isInitialLoading: false,
-      isLoading: false,
-      isLoadingError: false,
-      isPaused: false,
-      isPending: true,
-      isPlaceholderData: false,
-      isRefetchError: false,
-      isRefetching: false,
-      isStale: true,
-      isSuccess: false,
-      promise: Promise.resolve(undefined),
-      refetch: vi.fn(),
-      status: "pending",
-    } as never);
-  });
-
-  afterEach(() => {
-    for (const queryClient of queryClients.splice(0)) {
-      queryClient.clear();
-    }
-    cleanup();
-    useDashboardBentoStore.setState({
-      refreshToken: 0,
-      selectedTraceID: null,
-    });
-    useExportDialogStore.setState({
-      isExportDialogOpen: false,
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useFactoryTimelineStore.getState().reset();
-    resetSelectionHistoryStore();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
+  registerAppDashboardTestLifecycle();
 
   it("renders a trace drill-down for a selected work item", async () => {
     renderApp({
@@ -1212,33 +901,7 @@ describe("App current selection", () => {
 });
 
 describe("App current selection layout", () => {
-  beforeEach(() => {
-    window.localStorage.clear();
-    MockEventSource.instances = [];
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-  });
-
-  afterEach(() => {
-    for (const queryClient of queryClients.splice(0)) {
-      queryClient.clear();
-    }
-    cleanup();
-    useDashboardBentoStore.setState({
-      refreshToken: 0,
-      selectedTraceID: null,
-    });
-    useExportDialogStore.setState({
-      isExportDialogOpen: false,
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useFactoryTimelineStore.getState().reset();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
+  registerAppDashboardTestLifecycle({ resetSelectionHistory: false });
 
   it("keeps selection detail out of the workflow graph inspector layer", async () => {
     renderApp({
@@ -1780,33 +1443,7 @@ describe("App current selection layout", () => {
 });
 
 describe("App current selection terminal states", () => {
-  beforeEach(() => {
-    window.localStorage.clear();
-    MockEventSource.instances = [];
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-  });
-
-  afterEach(() => {
-    for (const queryClient of queryClients.splice(0)) {
-      queryClient.clear();
-    }
-    cleanup();
-    useDashboardBentoStore.setState({
-      refreshToken: 0,
-      selectedTraceID: null,
-    });
-    useExportDialogStore.setState({
-      isExportDialogOpen: false,
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useFactoryTimelineStore.getState().reset();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
+  registerAppDashboardTestLifecycle({ resetSelectionHistory: false });
 
   it("opens completed and failed work summaries and updates the trace card", async () => {
     renderApp({
