+++ ui/src/App.test.tsx
@@ -1,21 +1,23 @@
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import {
   act,
-  cleanup,
   fireEvent,
-  render,
   screen,
   waitFor,
   within,
 } from "@testing-library/react";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { App } from "./App";
+import { describe, expect, it, vi } from "vitest";
+import {
+  MockEventSource,
+  registerAppDashboardTestLifecycle,
+  removeTraceIDsFromSnapshot,
+  renderApp,
+  renderWithAppDashboardQueryClient,
+  requireValue,
+} from "./App.dashboard-test-harness";
 import type {
   DashboardRuntimeWorkstationRequest,
   DashboardSnapshot,
   DashboardTrace,
-  DashboardWorkItemRef,
-  DashboardWorkstationRequest,
 } from "./api/dashboard";
 import type { FactoryEvent } from "./api/events";
 import { FACTORY_EVENT_TYPES } from "./api/events";
@@ -37,7 +39,6 @@ import {
   scriptDashboardIntegrationFixtureIDs,
   scriptDashboardIntegrationTimelineEvents,
 } from "./components/dashboard/fixtures";
-import { installDashboardBrowserTestShims } from "./components/dashboard/test-browser-shims";
 import {
   semanticWorkflowDashboardSnapshot,
   twentyNodeDashboardSnapshot,
@@ -48,83 +49,15 @@ import {
   DASHBOARD_SUPPORTING_LABELS_CLASS,
 } from "./components/ui/dashboard-typography";
 import { formatDurationMillis } from "./components/ui/formatters";
-import { useDashboardBentoStore } from "./features/bento/state/dashboardBentoStore";
-import { reloadDashboardLayoutFromStorage } from "./features/bento/useDashboardLayout";
-import { useCurrentEditableFactoryDefinition } from "./features/current-factory-definition";
-import { resetSelectionHistoryStore } from "./features/current-selection/state/selectionHistoryStore";
-import {
-  createDefaultDashboardStreamState,
-  useDashboardStreamStore,
-} from "./features/dashboard/state/dashboardStreamStore";
+import { useDashboardStreamStore } from "./features/dashboard/state/dashboardStreamStore";
 import * as factoryPngExportModule from "./features/export/factory-png-export";
-import { useExportDialogStore } from "./features/export/state/exportDialogStore";
 import type { FactoryPngImportValue } from "./features/import";
 import * as factoryPngImportModule from "./features/import/factory-png-import";
-import type { WorldState } from "./features/timeline/state/factoryTimelineStore";
 import { useFactoryTimelineStore } from "./features/timeline/state/factoryTimelineStore";
 import {
   TraceDrilldownWidget,
   useTraceDrilldown,
 } from "./features/trace-drilldown";
-
-vi.mock("./features/current-factory-definition", async () => {
-  const actual = await vi.importActual("./features/current-factory-definition");
-
-  return {
-    ...actual,
-    useCurrentEditableFactoryDefinition: vi.fn(),
-  };
-});
-
-class MockEventSource {
-  public static instances: MockEventSource[] = [];
-
-  public onerror: ((event: Event) => void) | null = null;
-  public onopen: ((event: Event) => void) | null = null;
-
-  private readonly listeners = new Map<string, EventListener[]>();
-
-  constructor(public readonly url: string) {
-    MockEventSource.instances.push(this);
-  }
-
-  public addEventListener(type: string, listener: EventListener): void {
-    const existing = this.listeners.get(type) ?? [];
-    existing.push(listener);
-    this.listeners.set(type, existing);
-  }
-
-  public close(): void {}
-
-  public emit(type: string, data: unknown): void {
-    if (type === "snapshot") {
-      const state = useFactoryTimelineStore.getState();
-      const tracesByWorkID =
-        state.worldViewCache[state.selectedTick]?.tracesByWorkID ?? {};
-      seedTimelineSnapshot(data as DashboardSnapshot, tracesByWorkID);
-    }
-
-    const event = new MessageEvent(type, {
-      data: JSON.stringify(data),
-    });
-
-    for (const listener of this.listeners.get(type) ?? []) {
-      listener(event);
-    }
-  }
-}
-
-interface RenderAppOptions {
-  browserLanguage?: string | null;
-  browserLanguages?: readonly string[] | null;
-  initialLocale?: string | null;
-  locationSearch?: string | null;
-  snapshot: DashboardSnapshot;
-  timelineEvents?: FactoryEvent[];
-  timelineSnapshots?: DashboardSnapshot[];
-  traceFixtures?: Record<string, DashboardTrace>;
-  workstationRequestsByDispatchID?: Record<string, DashboardWorkstationRequest>;
-}
 
 function TraceDrilldownTestHarness({
   selectedWorkID,
@@ -403,14 +336,6 @@ function getWorkstationNodeByLabel(label: string): HTMLElement {
   return node;
 }
 
-function requireValue<T>(value: T | null | undefined, message: string): T {
-  if (value === null || value === undefined) {
-    throw new Error(message);
-  }
-
-  return value;
-}
-
 function expectFixedReviewWorkstationDimensions(): void {
   const reviewNode = getWorkstationNodeByLabel("Review");
 
@@ -564,73 +489,6 @@ function _expectRenderedWorkstationRequest(
       "The workstation request has not produced a response yet.",
     ),
   ).toBeTruthy();
-}
-
-function removeTraceIDFromWorkItem(
-  workItem: DashboardWorkItemRef,
-): DashboardWorkItemRef {
-  const withoutTraceID: DashboardWorkItemRef = { work_id: workItem.work_id };
-  if (workItem.display_name) {
-    withoutTraceID.display_name = workItem.display_name;
-  }
-  if (workItem.work_type_id) {
-    withoutTraceID.work_type_id = workItem.work_type_id;
-  }
-  return withoutTraceID;
-}
-
-function removeTraceIDsFromSnapshot(
-  snapshot: DashboardSnapshot,
-): DashboardSnapshot {
-  return {
-    ...snapshot,
-    runtime: {
-      ...snapshot.runtime,
-      active_executions_by_dispatch_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.active_executions_by_dispatch_id ?? {},
-        ).map(([dispatchID, execution]) => [
-          dispatchID,
-          {
-            ...execution,
-            trace_ids: [],
-            work_items: execution.work_items?.map(removeTraceIDFromWorkItem),
-          },
-        ]),
-      ),
-      current_work_items_by_place_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.current_work_items_by_place_id ?? {},
-        ).map(([placeID, workItems]) => [
-          placeID,
-          workItems.map(removeTraceIDFromWorkItem),
-        ]),
-      ),
-      session: {
-        ...snapshot.runtime.session,
-        provider_sessions: snapshot.runtime.session.provider_sessions?.map(
-          (attempt) => ({
-            ...attempt,
-            work_items: attempt.work_items?.map(removeTraceIDFromWorkItem),
-          }),
-        ),
-      },
-      workstation_activity_by_node_id: Object.fromEntries(
-        Object.entries(
-          snapshot.runtime.workstation_activity_by_node_id ?? {},
-        ).map(([nodeID, activity]) => [
-          nodeID,
-          {
-            ...activity,
-            active_work_items: activity.active_work_items?.map(
-              removeTraceIDFromWorkItem,
-            ),
-            trace_ids: [],
-          },
-        ]),
-      ),
-    },
-  };
 }
 
 function expectSeparatedStateMarkerZones(label: string, count: number): void {
@@ -1190,137 +1048,6 @@ const currentNamedFactoryExportResponse = {
     },
   ],
 } satisfies FactoryValue;
-const queryClients: QueryClient[] = [];
-let restoreBrowserTestShims: (() => void) | null = null;
-
-function timelineSnapshot(
-  snapshot: DashboardSnapshot,
-  tracesByWorkID: Record<string, DashboardTrace> = {},
-  workstationRequestsByDispatchID: Record<
-    string,
-    DashboardWorkstationRequest
-  > = {},
-): WorldState {
-  return {
-    ...snapshot,
-    relationsByWorkID: {},
-    tracesByWorkID,
-    workstationRequestsByDispatchID,
-    workRequestsByID: {},
-  };
-}
-
-function seedTimelineSnapshot(
-  snapshot: DashboardSnapshot,
-  tracesByWorkID: Record<string, DashboardTrace> = {},
-  workstationRequestsByDispatchID: Record<
-    string,
-    DashboardWorkstationRequest
-  > = {},
-): void {
-  useFactoryTimelineStore.setState({
-    events: [],
-    latestTick: snapshot.tick_count,
-    mode: "current",
-    receivedEventIDs: [],
-    selectedTick: snapshot.tick_count,
-    worldViewCache: {
-      [snapshot.tick_count]: timelineSnapshot(
-        snapshot,
-        tracesByWorkID,
-        workstationRequestsByDispatchID,
-      ),
-    },
-  });
-}
-
-function seedTimelineSnapshots(snapshots: DashboardSnapshot[]): void {
-  const worldViewCache = Object.fromEntries(
-    snapshots.map(
-      (snapshot) =>
-        [
-          snapshot.tick_count,
-          timelineSnapshot(snapshot) satisfies WorldState,
-        ] as const,
-    ),
-  );
-  const latestTick = Math.max(
-    ...snapshots.map((snapshot) => snapshot.tick_count),
-  );
-
-  useFactoryTimelineStore.setState({
-    events: [],
-    latestTick,
-    mode: "current",
-    receivedEventIDs: [],
-    selectedTick: latestTick,
-    worldViewCache,
-  });
-}
-
-function renderApp({
-  browserLanguage,
-  browserLanguages,
-  initialLocale,
-  locationSearch,
-  snapshot,
-  timelineEvents,
-  timelineSnapshots,
-  traceFixtures = {},
-  workstationRequestsByDispatchID = {},
-}: RenderAppOptions) {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        gcTime: Infinity,
-        retry: false,
-      },
-    },
-  });
-  queryClients.push(queryClient);
-
-  const fetchMock = vi
-    .fn()
-    .mockImplementation(async (input: RequestInfo | URL) => {
-      const path =
-        typeof input === "string"
-          ? input
-          : input instanceof URL
-            ? `${input.pathname}${input.search}`
-            : input.url;
-
-      throw new Error(`unexpected fetch for ${path}`);
-    });
-
-  vi.stubGlobal("fetch", fetchMock);
-  vi.stubGlobal("EventSource", MockEventSource);
-  reloadDashboardLayoutFromStorage();
-  if (timelineEvents) {
-    useFactoryTimelineStore.getState().replaceEvents(timelineEvents);
-  } else if (timelineSnapshots) {
-    seedTimelineSnapshots(timelineSnapshots);
-  } else {
-    seedTimelineSnapshot(
-      snapshot,
-      traceFixtures,
-      workstationRequestsByDispatchID,
-    );
-  }
-
-  const result = render(
-    <QueryClientProvider client={queryClient}>
-      <App
-        browserLanguage={browserLanguage}
-        browserLanguages={browserLanguages}
-        initialLocale={initialLocale}
-        locationSearch={locationSearch}
-      />
-    </QueryClientProvider>,
-  );
-
-  return { ...result, fetchMock };
-}
-
 function submitWorkCardControls() {
   const dashboardGrid = screen.getByRole("region", {
     name: "Infinite You bento board",
@@ -1465,21 +1192,10 @@ function renderTraceDrilldownHarness({
   selectedWorkID: string;
   timelineEvents: FactoryEvent[];
 }) {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        gcTime: Infinity,
-        retry: false,
-      },
-    },
-  });
-  queryClients.push(queryClient);
   useFactoryTimelineStore.getState().replaceEvents(timelineEvents);
 
-  return render(
-    <QueryClientProvider client={queryClient}>
-      <TraceDrilldownTestHarness selectedWorkID={selectedWorkID} />
-    </QueryClientProvider>,
+  return renderWithAppDashboardQueryClient(
+    <TraceDrilldownTestHarness selectedWorkID={selectedWorkID} />,
   );
 }
 
@@ -1569,92 +1285,6 @@ function _getDispatchHistoryCard(
   }
 
   return card;
-}
-
-function registerAppDashboardTestLifecycle(): void {
-  beforeEach(() => {
-    window.localStorage.clear();
-    MockEventSource.instances = [];
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-    resetSelectionHistoryStore();
-    vi.mocked(useCurrentEditableFactoryDefinition).mockReturnValue({
-      data: undefined,
-      error: null,
-      failureCount: 0,
-      failureReason: null,
-      fetchStatus: "idle",
-      isError: false,
-      isFetched: false,
-      isFetchedAfterMount: false,
-      isFetching: false,
-      isInitialLoading: false,
-      isLoading: false,
-      isLoadingError: false,
-      isPaused: false,
-      isPending: true,
-      isPlaceholderData: false,
-      isRefetchError: false,
-      isRefetching: false,
-      isStale: true,
-      isSuccess: false,
-      promise: Promise.resolve(undefined),
-      refetch: vi.fn(),
-      status: "pending",
-    } as never);
-  });
-
-  afterEach(() => {
-    for (const queryClient of queryClients.splice(0)) {
-      queryClient.clear();
-    }
-    cleanup();
-    useDashboardBentoStore.setState({
-      refreshToken: 0,
-      selectedTraceID: null,
-    });
-    useExportDialogStore.setState({
-      isExportDialogOpen: false,
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useFactoryTimelineStore.getState().reset();
-    resetSelectionHistoryStore();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
-}
-
-function registerAppFollowUpTestLifecycle(): void {
-  beforeEach(() => {
-    window.localStorage.clear();
-    MockEventSource.instances = [];
-    restoreBrowserTestShims = installDashboardBrowserTestShims();
-  });
-
-  afterEach(() => {
-    for (const queryClient of queryClients.splice(0)) {
-      queryClient.clear();
-    }
-    cleanup();
-    useDashboardBentoStore.setState({
-      refreshToken: 0,
-      selectedTraceID: null,
-    });
-    useExportDialogStore.setState({
-      isExportDialogOpen: false,
-    });
-    useDashboardStreamStore.setState({
-      streamState: createDefaultDashboardStreamState(),
-    });
-    useFactoryTimelineStore.getState().reset();
-    restoreBrowserTestShims?.();
-    restoreBrowserTestShims = null;
-    vi.restoreAllMocks();
-    vi.unstubAllGlobals();
-  });
 }
 
 describe("App shell import and export flows", () => {
@@ -3698,7 +3328,7 @@ describe("App dashboard layout and graph behavior", () => {
 });
 
 describe("App dashboard follow-up flows", () => {
-  registerAppFollowUpTestLifecycle();
+  registerAppDashboardTestLifecycle({ resetSelectionHistory: false });
 
   it("renders the submit-work card alongside the existing dashboard widgets", async () => {
     renderApp({ snapshot: terminalSnapshot });
